+++ jest.config.cjs
@@ -15,22 +15,11 @@ module.exports = {
     'node_modules',
     'Mock'
   ],
-  globals: {
-    'ts-jest': {
-      useESM: true
-    }
-  },
-  preset: 'ts-jest/presets/js-with-ts-esm',
   transform: {
-    '^.+\\.jsx?$': 'babel-jest',
     '^.+\\.tsx?$': 'ts-jest'
   },
-  transformIgnorePatterns: [
-    'node_modules/?!@jamashita'
-  ],
   moduleNameMapper: {
-    '^@jamashita/lluvia-(.*)$': '<rootDir>/packages/$1/src/index',
-    '^(\\.{1,2}/.*)\\.js$': '$1'
+    '^@jamashita/lluvia-(.*)$': '<rootDir>/packages/$1/src/index'
   },
   moduleFileExtensions: [
     'js',