+++ packages/tree/src/ClosureTable/Mock/MockClosureTableHierarchy.ts
@@ -1,5 +1,5 @@
-import { TreeID } from '../../TreeID.js';
-import { ClosureTableHierarchy } from '../ClosureTableHierarchy.js';
+import { TreeID } from '../../TreeID';
+import { ClosureTableHierarchy } from '../ClosureTableHierarchy';
 
 export class MockClosureTableHierarchy<K extends TreeID> extends ClosureTableHierarchy<K> {
   public constructor(ancestor: K, offspring: K) {