+++ packages/tree/src/ClosureTable/TreeIDFactory.ts
@@ -1,5 +1,5 @@
 import { Primitive } from '@jamashita/anden-type';
-import { TreeID } from '../TreeID.js';
+import { TreeID } from '../TreeID';
 
 export interface TreeIDFactory<K extends TreeID> {
   forge(id: Primitive): K;